+++ dags/oss_know/libs/base_dict/opensearch_index.py
@@ -16,7 +16,6 @@ OPENSEARCH_DISCOURSE_CATEGORY = "discourse_category"
 OPENSEARCH_DISCOURSE_TOPIC_LIST_OLD = "discourse_topic_list_old"
 OPENSEARCH_DISCOURSE_TOPIC_LIST = "discourse_topic_list"
 OPENSEARCH_DISCOURSE_TOPIC_CONTENT = "discourse_topic_content"
-# OPENSEARCH_DISCOURSE_TOPIC_CONTENT = "discourse_topic_content_tmp"
 OPENSEARCH_DISCOURSE_USER_LIST = "discourse_user"
 OPENSEARCH_DISCOURSE_USER_INFO = "discourse_user_info"
 OPENSEARCH_DISCOURSE_USER_ACTION = "discourse_user_action"
